refactor(routes): Refactors
- Refactored job APIs
